feat(web): workspace de organizacion /organizations/[id]/manage
Migra la pagina de detalle de organizacion (gestion de miembros) de
/panel/organizaciones/[id] a /organizations/[id]/manage, envuelta con
DashboardLayout. Redirect permanente desde la ruta legacy.
