+++ apps/web/src/app/panel/organizaciones/actions.ts
@@ -43,71 +43,5 @@ export async function createOrganizationAction(
   }
 
   revalidatePath('/panel/organizaciones');
-  redirect(`/panel/organizaciones/${data.id}`);
-}
-
-export async function addMemberAction(
-  orgId: string,
-  _prev: OrgActionResult,
-  formData: FormData,
-): Promise<OrgActionResult> {
-  const token = await requireSession(`/panel/organizaciones/${orgId}`);
-
-  const { t } = await getT();
-
-  const email = String(formData.get('email') ?? '').trim();
-  if (!email) {
-    return { status: 'error', message: t.organizaciones.err_email_required };
-  }
-
-  const { error, response } = await api.POST('/organizations/{id}/members', {
-    params: { path: { id: orgId } },
-    headers: authHeaders(token),
-    body: { email },
-  });
-
-  if (error !== undefined) {
-    if (response.status === 401) redirect(loginHref(`/panel/organizaciones/${orgId}`));
-    if (response.status === 403) {
-      return { status: 'error', message: t.organizaciones.err_owner_only };
-    }
-    if (response.status === 404) {
-      return { status: 'error', message: t.organizaciones.err_user_not_found };
-    }
-    if (response.status === 409) {
-      return { status: 'error', message: t.organizaciones.err_already_member };
-    }
-    return { status: 'error', message: t.organizaciones.err_add_member_failed };
-  }
-
-  revalidatePath(`/panel/organizaciones/${orgId}`);
-  return { status: 'success' };
-}
-
-export async function removeMemberAction(
-  orgId: string,
-  userId: string,
-): Promise<OrgActionResult> {
-  const token = await requireSession(`/panel/organizaciones/${orgId}`);
-
-  const { t } = await getT();
-
-  const { error, response } = await api.DELETE('/organizations/{id}/members/{userId}', {
-    params: { path: { id: orgId, userId } },
-    headers: authHeaders(token),
-  });
-
-  if (error !== undefined) {
-    if (response.status === 401) redirect(loginHref(`/panel/organizaciones/${orgId}`));
-    if (response.status === 403) {
-      return { status: 'error', message: t.organizaciones.err_owner_only };
-    }
-    if (response.status === 422) {
-      return { status: 'error', message: t.organizaciones.err_owner_cannot_remove_self };
-    }
-    return { status: 'error', message: t.organizaciones.err_remove_member_failed };
-  }
-
-  revalidatePath(`/panel/organizaciones/${orgId}`);
-  return { status: 'success' };
+  redirect(`/organizations/${data.id}/manage`);
 }
